# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Medical Admission API",
        "description": "API for managing medical patient admissions",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/obstetric-sheets": {
            "get": {
                "tags": [
                    "Obstetric Sheets"
                ],
                "summary": "Get all obstetric sheets",
                "operationId": "81ce437a5ea6eb40406edf4f147d7186",
                "responses": {
                    "200": {
                        "description": "Successful retrieval"
                    }
                }
            },
            "post": {
                "tags": [
                    "Obstetric Sheets"
                ],
                "summary": "Create a new obstetric sheet",
                "operationId": "64f45ac73f18c25133b75dfa6b347fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ObstetricSheet"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/obstetric-sheets/{id}": {
            "get": {
                "tags": [
                    "Obstetric Sheets"
                ],
                "summary": "Retrieve a single obstetric sheet",
                "operationId": "3e0affd96b2586629adc09585ae2aebf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Obstetric Sheets"
                ],
                "summary": "Update a specific obstetric sheet",
                "operationId": "1e01fbc132261ab54a49b1dbbcf1d63d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ObstetricSheet"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Obstetric Sheets"
                ],
                "summary": "Delete a specific obstetric sheet",
                "operationId": "92e7cf221361da6b45d4f02b8fa0f147",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ObstetricSheet": {
                "required": [
                    "patient_admission_id"
                ],
                "properties": {
                    "id": {
                        "description": "Obstetric sheet ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "patient_admission_id": {
                        "description": "Associated patient admission ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "reason_for_admission": {
                        "description": "Reason for admission",
                        "type": "string"
                    },
                    "admitting_impression": {
                        "description": "Admitting impression",
                        "type": "string"
                    },
                    "final_diagnosis": {
                        "description": "Final diagnosis",
                        "type": "string"
                    },
                    "pertinent_medical_history": {
                        "description": "Pertinent medical history",
                        "type": "string"
                    },
                    "educational_attainment": {
                        "description": "Educational attainment of the patient",
                        "type": "string"
                    },
                    "previous_pregnancies": {
                        "description": "History of previous pregnancies",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "lmp": {
                        "description": "Last menstrual period date",
                        "type": "string",
                        "format": "date"
                    },
                    "edc": {
                        "description": "Estimated date of confinement",
                        "type": "string",
                        "format": "date"
                    },
                    "aog": {
                        "description": "Age of gestation in weeks",
                        "type": "integer"
                    },
                    "weeks_pmp": {
                        "description": "Weeks post menstrual period",
                        "type": "integer"
                    },
                    "morning_sickness": {
                        "description": "Presence of morning sickness",
                        "type": "boolean"
                    },
                    "quickening": {
                        "description": "Quickening observations",
                        "type": "string"
                    },
                    "abnormal_signs": {
                        "description": "Abnormal signs noted",
                        "type": "string"
                    },
                    "primary_antenatal_condition": {
                        "description": "Primary antenatal condition",
                        "type": "string"
                    },
                    "antenatal_visits_first": {
                        "description": "Date of first antenatal visit",
                        "type": "string",
                        "format": "date"
                    },
                    "antenatal_visits_last": {
                        "description": "Date of last antenatal visit",
                        "type": "string",
                        "format": "date"
                    },
                    "contraceptive_methods": {
                        "description": "Contraceptive methods used",
                        "type": "string"
                    },
                    "additional_children_wanted": {
                        "description": "Additional children wanted by the patient",
                        "type": "string"
                    },
                    "history_of_present_illness": {
                        "description": "History of present illness",
                        "type": "string"
                    },
                    "general_condition": {
                        "description": "General condition of the patient",
                        "type": "string"
                    },
                    "bp": {
                        "description": "Blood pressure",
                        "type": "string"
                    },
                    "hr": {
                        "description": "Heart rate",
                        "type": "integer"
                    },
                    "rr": {
                        "description": "Respiratory rate",
                        "type": "integer"
                    },
                    "temp": {
                        "description": "Body temperature",
                        "type": "number",
                        "format": "float"
                    },
                    "weight": {
                        "description": "Patient's weight",
                        "type": "number",
                        "format": "float"
                    },
                    "height": {
                        "description": "Patient's height",
                        "type": "number",
                        "format": "float"
                    },
                    "fundic_height": {
                        "description": "Fundic height measurement",
                        "type": "string"
                    },
                    "presentation": {
                        "description": "Fetal presentation",
                        "type": "string"
                    },
                    "engaged": {
                        "description": "Whether the fetal head is engaged",
                        "type": "boolean"
                    },
                    "floating": {
                        "description": "Whether the fetal head is floating",
                        "type": "boolean"
                    },
                    "efw": {
                        "description": "Estimated fetal weight",
                        "type": "number",
                        "format": "float"
                    },
                    "fht": {
                        "description": "Fetal heart tone/rate",
                        "type": "integer"
                    },
                    "extremities": {
                        "description": "Condition of extremities",
                        "type": "string"
                    },
                    "edema": {
                        "description": "Presence and degree of edema",
                        "type": "string"
                    },
                    "albuminuria": {
                        "description": "Presence of albumin in urine",
                        "type": "boolean"
                    },
                    "glucosuria": {
                        "description": "Presence of glucose in urine",
                        "type": "boolean"
                    },
                    "hemoglobin": {
                        "description": "Hemoglobin level",
                        "type": "number",
                        "format": "float"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}